adds shift test
